# Improve performance of ASM's ClassReader.readStream()