@sbrannen 2021-03-11
The following is our general stance on modification of the repackaged ASM code.

> Please refrain from modifying classes under `org.springframework.asm`, `org.springframework.cglib`, and `org.springframework.objenesis`. Those include repackaged forks of the third-party libraries ASM, CGLIB, and Objenesis. Any refactoring to those classes should take place upstream in the originating repository. The Spring Framework will then pick up the changes when syncing with official updates of the forked third-party libraries.

In any case, let's see if @jhoeller wishes to patch our repackaged ASM code with your proposal.