@stsypanov 2021-03-11
@sbrannen Sorry, I totally forgot that `org.springframework.asm` is 3rd-party code