@sbrannen 2021-03-11
> @sbrannen Sorry, I totally forgot that `org.springframework.asm` is 3rd-party code

No problem.

I was about to point you to the ASM repository, but I see that you just created [this PR](https://gitlab.ow2.org/asm/asm/-/merge_requests/314) there 17 minutes ago. 😉 

@jhoeller, in light of that we can wait on that PR to get merged into ASM proper and then consume it later.